% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_script.R
\name{accuracy_plot}
\alias{accuracy_plot}
\title{Build a plot to help understand the accuracies}
\usage{
accuracy_plot(workflow_data, x_label, y_label, plot_title)
}
\arguments{
\item{workflow_data}{A data frame containing the workflow}

\item{x_label}{Name given to the x axis}

\item{y_label}{Name given to the y axis}

\item{plot_title}{Name given to the entire graph}
}
\value{
a visualization of the accuracy of the estimates with respect to
the number of neighbors
}
\description{
Creates a line plot depicting the relationship between two variables,
one of them being the accuracy
naming the axes and the plot with the given labels,
}
\examples{
test_data_2 <- data.frame(
 .metric = rep("accuracy", 30),
 neighbors = 1:30,
 mean = rnorm(30)
 )

accuracy_plot(test_data_2, "Neighbors", "Accuracy", "Accuracy Plot for Test Data 2")

}
