% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Create a grid of values to tune over}
\usage{
create_grid(min_neighbors, max_neighbors)
}
\arguments{
\item{min_neighbors}{An integer specifying the minimum number of neighbors to
include in the grid.}

\item{max_neighbors}{An integer specifying the maximum number of neighbors to
include in the grid.}
}
\value{
A tibble containing the grid of values to tune over.
}
\description{
This function generates a grid of values to tune over for the
number of nearest neighbors in the k-nearest neighbor classification model.
The grid includes integers ranging from the specified min_neighbors to
max_neighbors, inclusive.
}
\examples{
create_grid(1, 30)
}
