% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_knn_spec.R
\name{create_knn_spec}
\alias{create_knn_spec}
\title{Create a nearest neighbor specification object}
\usage{
create_knn_spec(weight_func)
}
\arguments{
\item{weight_func}{A string specifying the weight function to use.}
}
\value{
returns the specification object
}
\description{
This function creates a specification object for use in training k-nearest
neighbor models for classification.The \code{weight_func} parameter specifies the
weight function to use and the specification object is returned for use in a
modeling workflow.
}
\examples{
create_knn_spec("rectangular")

}
