% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_vfold.R
\name{create_vfold}
\alias{create_vfold}
\title{Create a v-fold cross-validation object}
\usage{
create_vfold(data, v, strata)
}
\arguments{
\item{data}{A dataframe containing the data.}

\item{v}{An integer specifying the number of folds.}

\item{strata}{A string specifying the name of the strata variable, if any.}
}
\value{
A v-fold cross-validation object for use in a workflow.
}
\description{
This function creates a v-fold cross-validation object for use in a modeling
workflow.The number of folds is specified by the \code{v} argument. If a \code{strata}
argument is given, then the cross-validation will be stratified by that
variable.
}
\examples{
create_vfold(drug_data, 5, "Cannabis")
}
