% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_script.R
\name{horizontal_hist}
\alias{horizontal_hist}
\title{Build a horizontal histogram that will help understand the pattern between
two variables}
\usage{
horizontal_hist(
  data,
  x_var,
  y_var,
  x_label,
  y_label,
  plot_title,
  plot_width,
  plot_height
)
}
\arguments{
\item{data}{A data frame containing}

\item{x_var}{The column that consists of the x values for the graph}

\item{y_var}{The column that consists of the y values for the graph}

\item{x_label}{Name given to the x axis}

\item{y_label}{Name given to the y axis}

\item{plot_title}{Name given to the entire graph}

\item{plot_width}{Width of the plot}

\item{plot_height}{Height of the plot}
}
\value{
a visualization of a horizontal histogram
}
\description{
Creates a horizontal histogram depicts relationship between 2 variables,
naming the axes and the plot with the given labels,
in a given size
}
\examples{
horizontal_hist(cannabis_and_nicotine, label, n, "",
  "Amount of participants", "Nicotine usage vs. Cannabis usage",
  plot_width = 8, plot_height = 6
)

}
