% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_dataset.R
\name{split_dataset}
\alias{split_dataset}
\title{Split the dataset into training and testing sets based on a strata variable}
\usage{
split_dataset(data_set, strata_variable, predictor)
}
\arguments{
\item{data_set}{the data set to be used for the algorithm}

\item{strata_variable}{a string representing the variable that will be set
as the strata in initial_split}

\item{predictor}{a string representing the variable to be used as the
predictor for the algorithm}
}
\value{
a list containing the training and the testing data sets
}
\description{
Splits the data into testing/training and ensures variables are appropriate
for classification algorithm
}
\examples{
data_set <- data.frame(
x = rnorm(100),
y = sample(c("A", "B", "C"), 100, replace = TRUE),
z = rnorm(100)
)
strata_variable <- "y"
predictor <- "x"

split_dataset(data_set, strata_variable, predictor)

}
