% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_wrangle_data.R
\name{wrangle_training_data}
\alias{wrangle_training_data}
\title{Clean and wrangle data}
\usage{
wrangle_training_data(training_data, predictor, strata_variable, group_labels)
}
\arguments{
\item{training_data}{a dataframe for training data}

\item{predictor}{a string representing the variable to be used as the
predictor for the algorithm}

\item{strata_variable}{a string representing the variable that will be set as
the strata in initial_split}

\item{group_labels}{list of strings representing the group names}
}
\value{
cleaned and wrangled training data grouped by predictor and
strata_variable & add a label column
}
\description{
This function creates a data frame by grouping training data by predictor
and strata_variables & adds a group_labels column
}
\examples{
training_data <- data.frame(
  predictor = c("yes", "no", "yes", "no"),
  strata_variable = c("yes", "yes", "no", "no"),
  outcome = c(1, 1, 0, 1)
  )

 predictor <- "predictor"
 strata_variable <- "strata_variable"
 group_labels <- c("Group 1")

 wrangle_training_data(
  training_data, strata_variable, predictor,
  group_labels)

}
