% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_recipe.R
\name{create_recipe}
\alias{create_recipe}
\title{Create a recipe for a dataframe}
\usage{
create_recipe(data, response_var)
}
\arguments{
\item{data}{A dataframe containing the data.}

\item{response_var}{A string specifying the name of the response variable.}
}
\value{
A recipe object for use in a workflow.
}
\description{
This function creates a recipe object for use in a modeling workflow.
The recipe includes a formula where the response variable is the
\code{response_var}and all other columns in the input dataframe are used as
predictors. The recipe object also includes steps for scaling and centering
the predictor variables.
}
\examples{
create_recipe(drug_data, "Cannabis")
}
