% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_vfold.R
\name{create_vfold}
\alias{create_vfold}
\title{Create a v-fold cross-validation object}
\usage{
create_vfold(data, v, strata)
}
\arguments{
\item{data}{A dataframe containing the data.}

\item{v}{An integer specifying the number of folds.}

\item{strata}{A string specifying the name of the strata variable, if any.}
}
\value{
A v-fold cross-validation object for use in a workflow.
}
\description{
This function creates a v-fold cross-validation object for use in a modeling
workflow.The number of folds is specified by the \code{v} argument. If a \code{strata}
argument is given, then the cross-validation will be stratified by that
variable.
}
\examples{
df <- data.frame(
Age = c(3, 5, 6, 7, 8, 9),
Gender = c(-0.95197, 0.48246, -0.48246, -0.48246, 0.48246, -0.48246),
Nscore = c(-0.14882, -0.67825, -0.46725, -1.32828, 0.62967, -0.24649),
Escore = c(-0.80615, -0.30033, -1.09207, 1.93886, 2.57309, 0.00332),
Oscore = c(-0.01928, -1.55521, -0.45174, -0.84732, -0.97631, -1.42424),
Ascore = c(0.59042, 2.03972, -0.30172, -0.30172, 0.76096, 0.59042),
Nicotine = c(0, 1, 1, 0, 1, 1),
Cannabis = c("no", "yes", "no", "yes", "no", "yes")
)

create_vfold(df, 5, "Cannabis")
}
