% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_drugs_workflow}
\alias{predict_drugs_workflow}
\title{Predicts the outcome of a target variable}
\usage{
predict_drugs_workflow(knn_wf, test_data = testing_data)
}
\arguments{
\item{knn_wf}{A fitted knn workflow object}

\item{test_data}{A dataframe containing test data to predict the outcome of
cannabis usage}
}
\value{
Writes the predicted outcome of cannabis usage for the new data in
the data folder
}
\description{
Given a fitted KNN workflow object and new test data, this function predicts
the outcome of a target variable for the test data using the KNN model.
The predicted outcome is then written to the data folder.
}
\examples{
predict_drugs_workflow(drugs_workflow, drugs_test)
}
